+++ neuron/src/app/Neuron/Web/Cache.hs
@@ -13,16 +13,16 @@ import Neuron.Zettelkasten.Graph.Type (ZettelGraph)
 import Neuron.Zettelkasten.ID (ZettelID)
 import Neuron.Zettelkasten.Zettel (ZettelError)
 import Relude
-import Rib.Shake
+import Rib.Shake (ribInputDir)
 import System.Directory (createDirectoryIfMissing)
-import System.FilePath
+import System.FilePath ((</>))
 
 data ReadMode
   = ReadMode_Direct Config
   | ReadMode_Cached
   deriving (Eq, Show)
 
-type CacheData = (ZettelGraph, Map ZettelID ZettelError)
+type CacheData = (ZettelGraph, Map ZettelID (NonEmpty ZettelError))
 
 cacheFile :: Action FilePath
 cacheFile = do