+++ neuron/test/Neuron/Zettelkasten/Query/ParserSpec.hs
@@ -13,7 +13,7 @@ import Data.Some (Some (Some))
 import Data.TagTree (Tag (Tag), mkTagPattern)
 import qualified Network.URI.Encode as E
 import Neuron.Zettelkasten.Connection (Connection (..))
-import Neuron.Zettelkasten.ID (unsafeMkZettelID)
+import Neuron.Zettelkasten.ID (ZettelID (ZettelID))
 import Neuron.Zettelkasten.Query.Parser (parseQueryLink)
 import Neuron.Zettelkasten.Zettel (ZettelQuery (..))
 import Relude
@@ -26,16 +26,16 @@ spec = do
   describe "z: links" $ do
     it "parses custom/hash ID" $ do
       parseQueryLink (asURI "z:/foo-bar")
-        `shouldBe` (Just $ Some $ ZettelQuery_ZettelByID (unsafeMkZettelID "foo-bar") OrdinaryConnection)
+        `shouldBe` (Just $ Some $ ZettelQuery_ZettelByID (ZettelID "foo-bar") OrdinaryConnection)
     it "supports branching links" $ do
       parseQueryLink (asURI "z:/foo-bar?type=branch")
-        `shouldBe` (Just $ Some $ ZettelQuery_ZettelByID (unsafeMkZettelID "foo-bar") Folgezettel)
+        `shouldBe` (Just $ Some $ ZettelQuery_ZettelByID (ZettelID "foo-bar") Folgezettel)
     it "parses prefixed short link" $ do
       parseQueryLink (asURI "z:/foo-bar")
-        `shouldBe` (Just $ Some $ ZettelQuery_ZettelByID (unsafeMkZettelID "foo-bar") OrdinaryConnection)
+        `shouldBe` (Just $ Some $ ZettelQuery_ZettelByID (ZettelID "foo-bar") OrdinaryConnection)
     it "resolves ambiguity using absolute URI" $ do
       parseQueryLink (asURI "z:/tags")
-        `shouldBe` (Just $ Some $ ZettelQuery_ZettelByID (unsafeMkZettelID "tags") OrdinaryConnection)
+        `shouldBe` (Just $ Some $ ZettelQuery_ZettelByID (ZettelID "tags") OrdinaryConnection)
     it "z:zettels" $ do
       parseQueryLink (asURI "z:zettels")
         `shouldBe` (Just $ Some $ ZettelQuery_ZettelsByTag [] OrdinaryConnection def)
@@ -64,12 +64,12 @@ spec = do
         `shouldBe` (Just $ Some $ ZettelQuery_TagZettel (Tag "日本語/おかもと先生/宿題"))
     it "z:/ i18n" $ do
       parseQueryLink (asURI $ "z:/" <> encodeUriPath "计算机")
-        `shouldBe` (Just $ Some $ ZettelQuery_ZettelByID (unsafeMkZettelID "计算机") OrdinaryConnection)
+        `shouldBe` (Just $ Some $ ZettelQuery_ZettelByID (ZettelID "计算机") OrdinaryConnection)
 
   describe "Regular Markdown links to zettels" $ do
     it "Supports full filename also" $ do
       parseQueryLink (asURI "foo-bar.md")
-        `shouldBe` (Just $ Some $ ZettelQuery_ZettelByID (unsafeMkZettelID "foo-bar") OrdinaryConnection)
+        `shouldBe` (Just $ Some $ ZettelQuery_ZettelByID (ZettelID "foo-bar") OrdinaryConnection)
 
   describe "Regular Markdown links (non-zettels)" $ do
     it "But lets normal links pass through" $ do