+++ CHANGELOG.md
@@ -10,12 +10,17 @@
   - Simplify error message UX for missing wiki-links (#448)
   - Baclinks surrounding context (#190)
   - Fix blockquote paragraph order in footnotes (#350)
+  - Allow specifying URL slugs in notes (#483)
 - Zettel format
   - Ignore punctuation in inline tags (#443)
   - Remove support for autolinks (`<..>`) (#449) (see [migration script](https://github.com/srid/neuron/issues/449#issuecomment-719062302))
     - Raw HTML now works without any special syntax.
   - Add Markdown highlighting extension (#453)
 - Unicode-aware in inline tags (#446)
+- CLI:
+  - `neuron open --id` is replaced by `neuron open --slug` (#483)
+- Removed features
+  - Aliases (#479)
 
 ## 1.0.1.0